refactor: Refactor s3 modules
