+++ lambda_backend/route_microservice/src/common/s3/index.ts
@@ -1,0 +1,1 @@
+export * from './s3';
