+++ lambda_backend/route_microservice/src/getRouteDetails.ts
@@ -6,7 +6,7 @@ import { getMiddlewareAddedHandler } from './common/middleware';
 import { getItemFromRouteTable } from './common/db';
 import { getRouteDetailsSchema } from './common/schema';
 import { GetRouteDetailsEvent, JwtPayload } from './common/types';
-import { restoreRouteURL } from './common/s3';
+import { restoreRouteURL } from './common/s3/utils';
 
 const getRouteDetails: Handler = async (event: GetRouteDetailsEvent) => {
   if (!process.env['ROUTE_TABLE_NAME'] || !process.env['COGNITO_USERPOOL_ID']) {
